refactor: better polyfill with `nodeIntegration:false` #24
